refactor: extract insight fund flow chart
